requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-filter==22.1
psycopg2==2.9.5
python-decouple==3.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
